+++ extensions/interactions/GraphInput/directives/graph-input-rules.service.spec.ts
@@ -18,8 +18,6 @@
 
 import { GraphInputRulesService } from
   'interactions/GraphInput/directives/graph-input-rules.service';
-import { GraphUtilsService } from
-  'interactions/GraphInput/directives/graph-utils.service';
 import { TestBed } from '@angular/core/testing';
 import { GraphAnswer } from 'interactions/answer-defs';
 
@@ -27,160 +25,6 @@ describe('Graph Input service', () => {
   let girs: GraphInputRulesService = null;
   beforeEach(() => {
     girs = TestBed.get(GraphInputRulesService);
-  });
-
-  describe('graph utilities', () => {
-    let utils: GraphUtilsService = null;
-    beforeEach(() => {
-      utils = TestBed.get(GraphUtilsService);
-    });
-
-    it('should construct an adjacency matrix from a graph', () => {
-      expect(utils.constructAdjacencyMatrix({
-        vertices: [{
-          label: 'a',
-          x: 1.0,
-          y: 1.0
-        }, {
-          label: 'b',
-          x: 2.0,
-          y: 2.0
-        }, {
-          label: 'c',
-          x: 3.0,
-          y: 3.0
-        }],
-        edges: [{
-          src: 0,
-          dst: 1,
-          weight: 1
-        }, {
-          src: 1,
-          dst: 2,
-          weight: 2
-        }],
-        isDirected: false,
-        isWeighted: true,
-        isLabeled: true
-      })).toEqual([
-        [null, 1, null],
-        [1, null, 2],
-        [null, 2, null]
-      ]);
-      expect(utils.constructAdjacencyMatrix({
-        vertices: [{
-          label: 'a',
-          x: 1.0,
-          y: 1.0
-        }, {
-          label: 'b',
-          x: 2.0,
-          y: 2.0
-        }, {
-          label: 'c',
-          x: 3.0,
-          y: 3.0
-        }],
-        edges: [{
-          src: 0,
-          dst: 1,
-          weight: 1
-        }, {
-          src: 1,
-          dst: 2,
-          weight: 2
-        }],
-        isDirected: false,
-        isWeighted: false,
-        isLabeled: true
-      })).toEqual([
-        [null, 1, null],
-        [1, null, 1],
-        [null, 1, null]
-      ]);
-    });
-
-    it('should find the next lexicographical permutation', () => {
-      let permutation: number[] = [0, 1, 2, 3];
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([0, 1, 3, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([0, 2, 1, 3]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([0, 2, 3, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([0, 3, 1, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([0, 3, 2, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 0, 2, 3]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 0, 3, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 2, 0, 3]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 2, 3, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 3, 0, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([1, 3, 2, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 0, 1, 3]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 0, 3, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 1, 0, 3]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 1, 3, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 3, 0, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([2, 3, 1, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 0, 1, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 0, 2, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 1, 0, 2]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 1, 2, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 2, 0, 1]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toEqual([3, 2, 1, 0]);
-      permutation = utils.nextPermutation(permutation);
-      expect(permutation).toBe(null);
-    });
-
-    it('should compare adjacency matrices with a permutation', () => {
-      expect(utils.areAdjacencyMatricesEqualWithPermutation([
-        [null, 1, 1],
-        [2, null, 1],
-        [1, 1, null]
-      ], [
-        [null, 1, 1],
-        [2, null, 1],
-        [1, 1, null]
-      ], [0, 1, 2])).toBe(true);
-      expect(utils.areAdjacencyMatricesEqualWithPermutation([
-        [null, 1, 1],
-        [2, null, 1],
-        [1, 1, null]
-      ], [
-        [null, 1, null],
-        [2, null, 1],
-        [1, 1, null]
-      ], [0, 1, 2])).toBe(false);
-      expect(utils.areAdjacencyMatricesEqualWithPermutation([
-        [null, 1, 2],
-        [2, null, 1],
-        [1, 1, null]
-      ], [
-        [null, 1, 1],
-        [2, null, 1],
-        [1, 2, null]
-      ], [2, 0, 1])).toBe(true);
-    });
   });
 
   let undirectedEmptyGraph = (): GraphAnswer => {
